Refactor: model 패키지를 domain과 service 패키지로 분리
