+++ src/main/java/lotto/domain/Money.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import java.util.Map;
 import lotto.constant.LottoConst;
