+++ src/main/java/lotto/view/OutputView.java
@@ -3,7 +3,7 @@ package lotto.view;
 import java.util.List;
 import java.util.Map;
 import lotto.constant.LottoConst;
-import lotto.model.Rank;
+import lotto.domain.Rank;
 
 public class OutputView {
