+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
