+++ src/test/java/lotto/domain/WinningNumbersTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
